add style and revise controller - DB populate

+++ src/controllers/videoController.js
@@ -1,3 +1,4 @@
+import User from "../models/User";
 import Video from "../models/Video"; // DB MODEL 사용
 
 /* 
@@ -21,7 +22,9 @@ import Video from "../models/Video"; // DB MODEL 사용
 */
 export const home = async (req, res) => {
   try {
-    const videos = await Video.find({});
+    const videos = await Video.find({})
+    .sort({ createAt: "desc"})
+    .populate("owner");
     return res.render("home", {pageTitle : "Home", videos });
   } catch {
     return res.status(404).render("server-error");
@@ -31,7 +34,10 @@ export const home = async (req, res) => {
 export const watch = async (req, res) => {
   // req 는 URL (/video/:id)에서 받은 변수정보를 담음.
   const { id } = req.params;
-  const video = await Video.findById(id);
+  // 비디오의 정보 + 해당 비디오의 작성자 닉네임 불러오기(ref 된 model 에서 가져올 정보)
+  const video = await Video.findById(id).populate("owner");
+  console.log(video);
+  
   if (video) {
     return res.render("watch", { pageTitle: video.title, video});
   } 
@@ -40,11 +46,19 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
   const { id } = req.params;
+  const { user: { _id }} = req.session;
   const video = await Video.findById(id);
-  if (video) {
-    return res.render("edit", {pageTitle: `Edit ${ video.title }`});
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video not found"})
   } 
-  return res.status(404).render("404", { pageTitle: "Video not found"})
+  /* 수정 시 작성자만 접속가능함 
+    video.owner : Object
+    _id : String
+  */
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect("/");
+  }
+  return res.render("edit", {pageTitle: `Edit ${ video.title }`});
   };
 
 // video 수정
@@ -79,30 +93,51 @@ export const getUpload = (req, res) => {
 };
 
 export const postUpload = async (req, res) => {
+  const { user : { _id } } = req.session;
   const { path: fileUrl } = req.file; // multer
   const { title, description, hashtags } = req.body;
 
   try {
-    await Video.create({ // creates Video data model / promise / save into DB
+    // 1. video model 에 data 넣기
+    const newVideo = await Video.create({ // creates Video data model / promise / save into DB
       title, // title : title
       description, // description : description
       fileUrl,
+      owner:_id,
       hashtags : hashtags.split(",").map((word)=> `#${word}`),
       meta: {
         views: 0,
         rating: 0,
       },
     })
+    // 2. User model 에 video 정보 추가하기
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save(); // -> 버그 해결(영상업로드시 해당 계정 비밀번호해싱발생)
+    return res.redirect("/");
   } catch (error) { // error 출력 error에서 _message 를 통해 간략한 에러사항을 알 수 있음.
     console.log(error);
     return res.status(400).render("upload", { pageTitle : "Upload Video", errorMesaage : error._message });
   }
-  return res.redirect("/");
+ 
 };
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
-  await Video.findByIdAndDelete(id); // delete
+  const { user: { _id }} = req.session;
+
+  const video = await Video.findById(id);
+  // 1. 해당 비디오가 존재하는가
+  if(!video) {
+    return res.status(404).render("404", {pageTitle: "Video not found."});
+  }
+  // 2. 지우고자하는 영상의 소유자인가
+  if (String(video.owner)!==String(_id)) {
+    return res.status(403).redirect("/");
+  }
+  
+  // delete
+  await Video.findByIdAndDelete(id);
   return res.redirect("/");
 };
 
@@ -114,9 +149,9 @@ export const search = async (req, res) => {
   if (keyword) { // keyword 가 입력되면 실행
       videos = await Video.find({
         title: {
-          $regex: new RegExp(keyword, "i"), // 정규식 사용 할 수 있게 해줌
+          $regex: new RegExp(`${keyword}$`, "i"), // 정규식 사용 할 수 있게 해줌
       }, // where title = keyword
-    });
+    }).populate("owner");
   }
   return res.render("search", { pageTitle: "Search", videos});
 }
